feat(alias): Move parser to separate package
