+++ src/main/java/seedu/address/logic/parser/alias/AliasCommandParser.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.alias;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
@@ -7,6 +7,10 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_COMMAND;
 
 import seedu.address.commons.core.Alias;
 import seedu.address.logic.commands.alias.AliasCommand;
+import seedu.address.logic.parser.ArgumentMultimap;
+import seedu.address.logic.parser.ArgumentTokenizer;
+import seedu.address.logic.parser.Parser;
+import seedu.address.logic.parser.ParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 public class AliasCommandParser implements Parser<AliasCommand> {
